Setup initial `cargo-vet` crate auditing (#141)
Adds initial configurations for our crate auditing workflow with
[`cargo-vet`](https://github.com/mozilla/cargo-vet). This just adds the
standard files and configures our imported registries as well as default
trusted publishers.

Want to experiment with us using this auditing workflow in this public
library crate as we've only been testing it so far on our larger private
monorepo. This imports the audits from that through
https://github.com/EmbarkStudios/rust-ecosystem/blob/main/audits.toml.