Merge packed and unpacked into single type
This commit consolidates the packed and unpacked data from frame_data
into a single enum. The goal of this change is to minimize the number of
locks required when retrieving information about a frame.

Previously, the process of reading the packing information for stats
required four separate locks:
 1. One to check 'has_unpacked'
 2. Another to verify 'has_unpacked' within 'unpacked_size'
 3. One to retrieve 'unpacked_size'
 4. And finally, one to retrieve 'packed_size'

With this optimization, the lock count is reduced to a single one,
achieved through invoking the 'packing_info' function.