Remove `imgui` support in favor of `egui` (#140)
Most Rust project use the native `egui` instead, including us so we
don't want to maintain `imgui` support that we don't use.

Significantly reduces the dependency tree for development as we were
pulling in old `winit` and almost deprecated `glium` renderers for it
(`egui` uses `wgpu`).

cargo-vet stats (the exempted number is the most important):
- Before: `Vetting Succeeded (168 fully audited, 10 partially audited,
173 exempted)`
- After: `Vetting Succeeded (151 fully audited, 10 partially audited,
133 exempted)`

Resolves: #139
Resolves: #134