Remove `instant` dependency when not building for web
This dependency was introduced originally in #112.

It is a platform abstraction, but it is a tiny one and it is not harder to just use the std type which avoids always pulling in `instant` including for `wasm32-unknown-unknown` when not building for the web where it has no purpose at all.

Want to have as few dependencies as possible on core low level crates like `puffin` that almost all of our code depends on.