Remove file I/O in main `puffin` crate
The `puffin` crate is intended to be tiny and the main interface that all code depends on, but it had utilities to save the profile view to a file which adds code that depends on std file system APIs unnecessarily.

This generally doesn't matter to much, but we want to strive forward tighter capabilities in our dependencies and tools like [`cackle`](https://github.com/davidlattimore/cackle) detected that the `puffin` crate was doing file I/O and required that users allow that.

So this removes that support and requires users (mainly puffin viewer app) to handle file saving/loading it self, which is more appropriate.