🪂 Disable `drag_to_scroll` on frame selection `ScrollArea` (#214)
Redo of #207

### Checklist

* [x] I have read the [Contributor Guide](../CONTRIBUTING.md)
* [x] I have read and agree to the [Code of
Conduct](../CODE_OF_CONDUCT.md)
* [x] I have added a description of my changes and why I'd like them
included in the section below

### Description of Changes
In the version `0.27` release of `egui`, the interaction handling was
changed. This caused the scope selection in the flamegraph to be
unselectable. By disabling the `drag_to_scroll` on the related
`ScrollArea` this issue is solved.

The latest scope selection does not suffer from this, as it does not use
a `ScrollArea`.

Note that this does not have any behavioral effect, but instead results
in the same behavior of click/drag/scroll when on egui `0.26`!

### Related Issues
- Fixes #205

```[tasklist]
### Tested on
- [x] Windows
- [x] Linux
- [x] Mac
- [x] Android
```